requirements.txt
groq
chromadb
python-dotenv
pytest